Separate the classroom's rules from the code that applies them
hidden_from was in the same file as student_can_read, and it is not a rule --
it walks every class and every text and applies one. Mixing the two means the
file is no longer readable in a minute, because you have to sort the policy
from the traversal yourself.

classroom/rules.py now holds only policies: functions that answer "is this
allowed?" from values they are given, with no traversal, no query and no I/O.
Two of them, 44 lines including the prose:

    student_can_read(student, text)
    app_is_reduced_to_classroom(user, classes)

Note the second takes the classes rather than fetching them -- that is what
keeps it a rule.

classroom/queries.py finds things and applies those: classes_of, texts_of,
classroom_of, hidden_from, sees_only_class_texts. Queries call rules; rules
never call queries.

No behaviour change: 298 tests and 20 subtests pass unchanged.
