+++ zeeguu/core/classroom/__init__.py
@@ -1,0 +1,18 @@
+from zeeguu.core.classroom.queries import (
+    classes_of,
+    classroom_of,
+    hidden_from,
+    sees_only_class_texts,
+    texts_of,
+)
+from zeeguu.core.classroom.rules import app_is_reduced_to_classroom, student_can_read
+
+__all__ = [
+    "app_is_reduced_to_classroom",
+    "classes_of",
+    "classroom_of",
+    "hidden_from",
+    "sees_only_class_texts",
+    "student_can_read",
+    "texts_of",
+]
